Changing the output of which template you're using to be shown at the bottom of the page instead of the top
When you're logged into the admin area and go to a page on the site we
have a function that shows you which template is being applied.  In the
past we showed that on the top of the page but in some circumstances
(like when you have a fixed element on the top of the page) it gets
hidden and causes other elements to get pushed down the page in weird
ways.  This new way will fix that you just need to skip to the bottom
of the page to see it.